notes v1
Patrick printed the box for me and I assembled it. In principle it works (Video). I would change the following in the next iteration:
The holding above the motor should be 1mm higher
The distance between front and back should be 4mm longer
In the almost middle (1mm translated towards the back) could be 5mm high "bowl" designed to hold a 52mm sphere (golf ball)
The display should point to the other side
the gap for the display cables should be 5mm wider in total
The lid can not be fully opened. make it 1mm shorter towards the back (if we keep these joints)
two 2mm holes all the way through the lid to fix the LED. Their centers 3mm from the front and 2mm from the sides
the thing on the lid is 1mm to thick and does not fit into the stick.
The stick design blocks smaller angles between rotor and stick (photo).